+++ src/Adapter/Filesystem/Transaction.php
@@ -48,12 +48,14 @@ final class Transaction implements TransactionInterface
     /**
      * @template R
      *
-     * @return callable(R): Attempt<R>
+     * @param R $value
+     *
+     * @return Attempt<R>
      */
     #[\Override]
-    public function commit(): callable
+    public function commit(mixed $value): Attempt
     {
-        return fn(mixed $value) => $this
+        return $this
             ->notCommitted
             ->root()
             ->all()
@@ -66,16 +68,16 @@ final class Transaction implements TransactionInterface
     /**
      * @template R
      *
-     * @return callable(R): Attempt<R>
+     * @param R $value
+     *
+     * @return Attempt<R>
      */
     #[\Override]
-    public function rollback(): callable
+    public function rollback(mixed $value): Attempt
     {
-        return function(mixed $value) {
-            $this->notCommitted = $this->reset();
+        $this->notCommitted = $this->reset();
 
-            return Attempt::result($value);
-        };
+        return Attempt::result($value);
     }
 
     /**